% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Santiago-package.R
\docType{package}
\name{Santiago-package}
\alias{Santiago}
\alias{Santiago-package}
\title{Santiago: A package based on a paper-based survey conducted face-to-face in Santiago in 2016}
\description{
A dataset containing information about socio-demographics, health-related, perceptions and travel behaviour, travel choices and planning, social interaction factors, built environment of users of active and motorized modes of transportation.

The goal of Santiago is to contribute to psychological impacts on travelers by examining the feelings of stress by users of active and motorized modes of transportation.Furthermore, it also investigates the importance that travelers attach to their feelings of stress.
}
\references{
\url{https://github.com/NiloofarNL/Santiago}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NiloofarNL/Santiago}
  \item \url{https://niloofarnl.github.io/Santiago}
  \item Report bugs at \url{https://github.com/NiloofarNL/Santiago/issues}
}

}
\author{
\strong{Maintainer}: Niloofar Nalaee \email{nalaeen@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Niloofar Nalaee, School of Earth, Environment and Society, McMaster University \email{nalaeen@mcmaster.ca}
}
\keyword{internal}
